+++ src/main/scala/diy_zio/Main.scala
@@ -4,16 +4,6 @@ package diy_zio
 //import zio.*
 import diy.zio.*
 
-//object Main extends App:
-//  Runtime.default.unsafeRunSync(program)
-//
-//  lazy val program =
-//    for
-//      _    <- console.putStrLn("What is your name?")
-//      name <- ZIO.succeed("Julian") // console.getStrLn
-//      _    <- console.putStrLn(s"Hello $name")
-//    yield ()
-
 object businessLogic:
   type BusinessLogic = Has[BusinessLogic.Service]
 
@@ -21,50 +11,68 @@ object businessLogic:
     trait Service:
       def evenPicturesOf(topic: String): ZIO[Any, Nothing, Boolean]
 
-    lazy val live: ZIO[Google, Nothing, Service] =
-      ZIO.fromFunction(make)
-    def make(google: Google): Service =
-      (topic: String) => google.countPicturesOf(topic).map(_ % 2 == 0)
+    lazy val live: ZLayer[google.Google, Nothing, BusinessLogic] =
+      ZLayer.fromService(make)
+    def make(_google: google.Google.Service): Service =
+      topic => _google.countPicturesOf(topic).map(_ % 2 == 0)
 
   def evenPicturesOf(topic: String): ZIO[BusinessLogic, Nothing, Boolean] =
     ZIO.accessM(_.get.evenPicturesOf(topic))
 
-trait Google:
-  def countPicturesOf(topic: String): ZIO[Any, Nothing, Int]
+object google:
+  type Google = Has[Google.Service]
+  object Google:
+    trait Service:
+      def countPicturesOf(topic: String): ZIO[Any, Nothing, Int]
+  def countPicturesOf(topic: String): ZIO[Google, Nothing, Int] =
+    ZIO.accessM(_.get.countPicturesOf(topic))
 
 object GoogleImpl:
-  lazy val live: ZIO[Any, Nothing, Google] = ZIO.succeed(make)
-  lazy val make: Google = topic => ZIO.succeed(if topic == "cats" then 27193 else 183208)
+  lazy val live: ZLayer[Any, Nothing, google.Google] = ZLayer.succeed(make)
+  def make: google.Google.Service = topic => ZIO.succeed(if topic == "cats" then 27193 else 183208)
+
+object controller:
+  type Controller = Has[Controller.Service]
+  object Controller:
+    trait Service:
+      def run: ZIO[Any, Nothing, Unit]
+
+    lazy val live: ZLayer[businessLogic.BusinessLogic & console.Console, Nothing, Controller] =
+      ZLayer.fromServices(make)
+
+    def make(bl: businessLogic.BusinessLogic.Service, c: console.Console.Service): Service =
+      new:
+        lazy val run: ZIO[Any, Nothing, Unit] =
+          for
+            cats <- bl.evenPicturesOf("cats")
+            _    <- c.putStrLn(cats.toString)
+            dogs <- bl.evenPicturesOf("dogs")
+            _    <- c.putStrLn(dogs.toString)
+          yield ()
+
+  lazy val run: ZIO[Controller, Nothing, Unit] =
+    ZIO.accessM(_.get.run)
 
 object DependencyGraph:
-  lazy val live: ZIO[Any, Nothing, businessLogic.BusinessLogic.Service] =
+  lazy val live: ZLayer[Any, Nothing, controller.Controller] =
     for
-      googleMaker        <- GoogleImpl.live
-      businessLogicMaker <- businessLogic.BusinessLogic.live.provide(googleMaker)
-    yield businessLogicMaker
+      (googleMaker, consoleMaker) <- GoogleImpl.live.zip(console.Console.live)
+      businessLogicMaker          <- businessLogic.BusinessLogic.live.provide(googleMaker)
+      controllerMaker <- controller.Controller.live.provide(
+        businessLogicMaker `union` consoleMaker,
+      )
+    yield controllerMaker
 
-  lazy val make: businessLogic.BusinessLogic.Service =
-    val _google        = GoogleImpl.make
-    val _businessLogic = businessLogic.BusinessLogic.make(_google)
-    _businessLogic
+  lazy val make: controller.Controller.Service =
+    val (_google, _console) = (GoogleImpl.make, console.Console.make)
+    val _businessLogic      = businessLogic.BusinessLogic.make(_google)
+    val _controller         = controller.Controller.make(_businessLogic, _console)
+    _controller
 
 object Main extends scala.App:
-//  println("Hello World!")
-////  Runtime.default.unsafeRunSync(program.provide(DependencyGraph.make))
   Runtime.default.unsafeRunSync(program)
 
   lazy val program =
-    for
-      logic <- DependencyGraph.live
-      p     <- makeProgram.provideSome[ZEnv](_ `union` Has(logic))
-    yield p
-
-//  def makeProgram(businessLogic: BusinessLogic) =
-  lazy val makeProgram =
-    for
-      env  <- ZIO.environment[console.Console & businessLogic.BusinessLogic]
-      cats <- businessLogic.evenPicturesOf("cats")
-      _    <- console.putStrLn(cats.toString)
-      dogs <- businessLogic.evenPicturesOf("dogs")
-      _    <- console.putStrLn(dogs.toString)
-    yield ()
+    // DependencyGraph.live.zio.flatMap(_.get.run)
+    // DependencyGraph.live.zio.flatMap(r => controller.run.provide(r))
+    controller.run.provideLayer(DependencyGraph.live)