+++ CHANGELOG.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
 
+## [0.7.0] - 2022-08-03
+
+### Added
+- `ZLayer`
+- `ZIO.zip`
+- `ZIO.provideLayer`
+- `ZEnv.live`
+
+### Changed
+- Made `console.Console` & `Runtime` use `ZLayer`
+- Rewrote `Main` to use `ZLayer`
+
+
 ## [0.6.0] - 2022-08-02
 
 ### Added
@@ -67,7 +80,8 @@ and this project adheres to [Semantic Versioning].
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
 <!-- Versions -->
-[Unreleased]: https://github.com/jaacko-torus/diy-zio/compare/v0.6.0...HEAD
+[Unreleased]: https://github.com/jaacko-torus/diy-zio/compare/v0.7.0...HEAD
+[0.7.0]: https://github.com/jaacko-torus/diy-zio/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/jaacko-torus/diy-zio/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/jaacko-torus/diy-zio/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/jaacko-torus/diy-zio/compare/v0.3.0...v0.4.0